+++ includes/globals.php
@@ -5,7 +5,7 @@
  */
 
 // Current CMS version
-const VERSION = '1.7.4';
+const VERSION = '1.7.5';
 
 /**
  * Display copyright on the admin dashboard.